refactor: app 디렉토리의 css 폴더 제거 및 최상위 styles 폴더 생성
